controller now in state where doesnt crash world when placed. TESR now renders for controller

+++ src/main/java/org/atomicworkshop/jammachine/sequencing/SequencerSet.java
@@ -36,7 +36,7 @@ public class SequencerSet implements Iterable<Sequencer>
 		}
 	}
 
-	public long getBeatsPerMinute()
+	public int getBeatsPerMinute()
 	{
 		return beatsPerMinute;
 	}
